refactor: 패키지명 규칙 통일
- 자바 패키지명은 전부 소문자가 관례입니다
