+++ src/main/java/com/findex/mapper/AutoSyncConfigMapper.java
@@ -1,6 +1,6 @@
 package com.findex.mapper;
 
-import com.findex.dto.autoSyncConfig.AutoSyncConfigDto;
+import com.findex.dto.autosyncconfig.AutoSyncConfigDto;
 import com.findex.entity.AutoSyncConfig;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
