+++ src/main/java/com/findex/repository/autosyncconfig/AutoSyncConfigQueryRepository.java
@@ -1,6 +1,6 @@
-package com.findex.repository.autoSyncRepository;
+package com.findex.repository.autosyncconfig;
 
-import com.findex.dto.autoSyncConfig.AutoSyncConfigQuery;
+import com.findex.dto.autosyncconfig.AutoSyncConfigQuery;
 import com.findex.dto.response.CursorPageResponse;
 
 public interface AutoSyncConfigQueryRepository {
