+++ src/main/java/com/findex/service/AutoSyncConfigService.java
@@ -1,13 +1,13 @@
 package com.findex.service;
 
-import com.findex.dto.autoSyncConfig.AutoSyncConfigDto;
-import com.findex.dto.autoSyncConfig.AutoSyncConfigQuery;
-import com.findex.dto.autoSyncConfig.AutoSyncConfigUpdateRequest;
+import com.findex.dto.autosyncconfig.AutoSyncConfigDto;
+import com.findex.dto.autosyncconfig.AutoSyncConfigQuery;
+import com.findex.dto.autosyncconfig.AutoSyncConfigUpdateRequest;
 import com.findex.dto.response.CursorPageResponse;
 import com.findex.entity.AutoSyncConfig;
 import com.findex.exception.NotFoundException;
 import com.findex.mapper.AutoSyncConfigMapper;
-import com.findex.repository.autoSyncRepository.AutoSyncConfigRepository;
+import com.findex.repository.autosyncconfig.AutoSyncConfigRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
